Love-7/LoveS-task7-task-202411081201/gen_src/TextAnalyzer.java
// TextAnalyzer.java
import java.util.List;
import java.util.Map;

public class TextAnalyzer {
    private FileWordReader reader;
    private WordCounter counter;

    public TextAnalyzer(String fileName) {
        // Constructor implementation
    }

    public int totalWords() {
        // Method implementation
    }

    public List<Map.Entry<String, Integer>> topFrequentWords(int n) {
        // Method implementation
    }

    public int uniqueWordCount() {
        // Method implementation
    }
}

// FileWordReader.java
public class FileWordReader {
    public FileWordReader(String fileName) {
        // Constructor implementation
    }

    public List<String> getWords() {
        // Method implementation
    }
}

// WordCounter.java
import java.util.Map;

public class WordCounter {
    public void countWords(List<String> words) {
        // Method implementation
    }

    public Map<String, Integer> getWordCounts() {
        // Method implementation
    }
}